Added complete buttons
buttons are color coded for all.  Manager shall be able to mark task,
subtask and milestones complete.  Once user clicks on any item it will
show complete if it is complete or it will show mark complete if it is
not complete.  Buttons are color coded: Red for not complete and Green
for complete!